pyproject.toml
[tool.poetry]
name = "hack104_rec"
version = "0.1.0"
description = ""
authors = ["顏孜羲 <joseph.yen@gmail.com>"]

[tool.poetry.dependencies]
python = "*"
carriage = "^0.4.4"
ipykernel = "^4.8"
# pyspark = "^2.3"
findspark = "^1.3"
jieba = "^0.39.0"
opencc-python-reimplemented = "^0.1.3"
pandas = "^0.23.1"
autoflake = "^1.2"
flake8 = "^3.5"
isort = "^4.3"
autopep8 = "^1.3"
pyspark = "=2.3.1"
icecream = "^1.3"
pudb = "^2018.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
